# FWC.RMS.WebApi/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Revenue Management System",
    "description" : "Florida Fish and Wildlife Conservation Commision - Revenue Management System.",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "apiteam@fwc.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "https://fws.swagger.io/v1"
  }, {
    "url" : "http://fws.swagger.io/v1"
  } ],
...
  "paths" : {
    "/departmentDocumentSearch" : {
      "get" : {
        "tags" : [ "departmentDocumentSearch" ],
        "summary" : "Search Department Document records by keyword",
        "operationId" : "searchDepartmentDocumentsByKeyword",
        "parameters" : [ {
          "name" : "keyword",
          "in" : "query",
          "description" : "Search Keyword (FirstName or LastName)",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DepartmentDocumentSearchResponse"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      },
      "post" : {
        "tags" : [ "departmentDocumentSearch" ],
        "summary" : "Search Department Document records",
        "operationId" : "searchDepartmentDocuments",
        "requestBody" : {
          "description" : "Department Document search request",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DepartmentDocumentSearchRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DepartmentDocumentSearchResponse"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      }
    },
...
    "/transmittals/{transmittalNumber}/departmentDocuments" : {
      "get" : {
        "tags" : [ "departmentDocument" ],
        "summary" : "get Department Document records",
        "operationId" : "getDepartmentDocuments",
        "parameters" : [ {
          "name" : "transmittalNumber",
          "in" : "path",
          "description" : "Transmittal Number",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DepartmentDocument"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      },
      "post" : {
        "tags" : [ "departmentDocument" ],
        "summary" : "Create a Department Document record",
        "operationId" : "createDepartmentDocument",
        "parameters" : [ {
          "name" : "transmittalNumber",
          "in" : "path",
          "description" : "Transmittal Number",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "Department Document request payload",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DepartmentDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepartmentDocument"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      }
    },
    "/transmittals/{transmittalNumber}/departmentDocuments/{departmentDocumentsNumber}" : {
      "patch" : {
        "tags" : [ "departmentDocument" ],
        "summary" : "Update a Department Document record",
        "operationId" : "updateDepartmentDocument",
        "parameters" : [ {
          "name" : "transmittalNumber",
          "in" : "path",
          "description" : "Transmittal Number",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "departmentDocumentsNumber",
          "in" : "path",
          "description" : "Department Document Number",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "DepartmentDocument request payload",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DepartmentDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepartmentDocument"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "DepartmentDocumentSearchRequest" : {
        "required" : [ "firstName" ],
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "companyName" : {
            "type" : "string"
          },
          "departmentDocumentNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ddnSearchType" : {
            "type" : "string"
          },
          "checkNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "depositNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transmittalNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transmittalStatus" : {
            "type" : "string"
          },
          "checkAmount" : {
            "type" : "number",
            "format" : "double"
          },
          "cashListing" : {
            "type" : "string"
          },
          "eoCode" : {
            "type" : "string"
          },
          "objectCode" : {
            "type" : "string"
          },
          "depositDateFrom" : {
            "type" : "string",
            "format" : "date"
          },
          "depositDateTo" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "DepartmentDocumentSearchResponse" : {
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "companyName" : {
            "type" : "string"
          },
          "departmentDocumentNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "checkNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "checkAmount" : {
            "type" : "number",
            "format" : "double"
          },
          "checkDate" : {
            "type" : "string",
            "format" : "date"
          },
          "transmittalNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transmittalStatus" : {
            "type" : "string"
          },
          "depositNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cashListing" : {
            "type" : "string"
          }
        },
        "example" : {
          "firstName" : "firstName",
          "lastName" : "lastName",
          "checkNumber" : 6,
          "depositNumber" : 5,
          "checkAmount" : 1.4658129805029452,
          "companyName" : "companyName",
          "transmittalStatus" : "transmittalStatus",
          "departmentDocumentNumber" : 0,
          "checkDate" : "2000-01-23",
          "transmittalNumber" : 5,
          "cashListing" : "cashListing"
        }
      },
...
      "DepartmentDocumentRequest" : {
        "type" : "object",
        "properties" : {
          "checkNumber" : {
            "type" : "integer"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "companyName" : {
            "type" : "string"
          },
          "checkAmount" : {
            "type" : "number",
            "format" : "double"
          },
          "cashListing" : {
            "type" : "string"
          },
          "Comments" : {
            "type" : "string"
          }
        }
      },
      "DepartmentDocument" : {
        "required" : [ "departmentDocumentNumber", "transmittalNumber" ],
        "type" : "object",
        "properties" : {
          "departmentDocumentNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transmittalNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "checkNumber" : {
            "type" : "integer"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "companyName" : {
            "type" : "string"
          },
          "checkAmount" : {
            "type" : "number",
            "format" : "double"
          },
          "cashListing" : {
            "type" : "string"
          },
          "comments" : {
            "type" : "string"
          },
          "dateTimeStamp" : {
            "type" : "string",
            "format" : "datetime"
          }
        },
        "example" : {
          "firstName" : "firstName",
          "lastName" : "lastName",
          "checkNumber" : 1,
          "comments" : "comments",
          "checkAmount" : 5.962133916683182,
          "companyName" : "companyName",
          "dateTimeStamp" : "dateTimeStamp",
          "departmentDocumentNumber" : 0,
          "transmittalNumber" : 6,
          "cashListing" : "cashListing"
        }
      }
    },
    "securitySchemes" : {
      "fws_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "http://fws.swagger.io/oauth/dialog",
            "scopes" : {
              "write:transmittals" : "modify transmittals in your account",
              "read:transmittals" : "read your transmittals"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}